requirements.txt
fastapi
uvicorn
pyparsing
starlette
Pillow
numpy
python-jose
passlib
python-multipart